Basics readt

+++ app/src/main/java/techgravy/sunshine/models/Coord.java
@@ -9,47 +9,56 @@ import co.uk.rushorm.core.RushCore;
 /**
  * Created by aditlal on 13/04/16.
  */
-public class Coord implements Rush{
+public class Coord implements Rush {
     @SerializedName("lon")
-    private double lon;
+    private float lon;
 
     @SerializedName("lat")
-    private double lat;
+    private float lat;
 
-    public double getLon ()
-    {
+    public float getLon() {
         return lon;
     }
 
-    public void setLon (double lon)
-    {
+    public void setLon(float lon) {
         this.lon = lon;
     }
 
-    public double getLat ()
-    {
+    public float getLat() {
         return lat;
     }
 
-    public void setLat (double lat)
-    {
+    public void setLat(float lat) {
         this.lat = lat;
     }
 
     @Override
-    public String toString()
-    {
-        return "ClassPojo [lon = "+lon+", lat = "+lat+"]";
+    public String toString() {
+        return "ClassPojo [lon = " + lon + ", lat = " + lat + "]";
     }
 
     @Override
-    public void save() { RushCore.getInstance().save(this); }
+    public void save() {
+        RushCore.getInstance().save(this);
+    }
+
     @Override
-    public void save(RushCallback callback) { RushCore.getInstance().save(this, callback); }
+    public void save(RushCallback callback) {
+        RushCore.getInstance().save(this, callback);
+    }
+
     @Override
-    public void delete() { RushCore.getInstance().delete(this); }
+    public void delete() {
+        RushCore.getInstance().delete(this);
+    }
+
     @Override
-    public void delete(RushCallback callback) { RushCore.getInstance().delete(this, callback); }
+    public void delete(RushCallback callback) {
+        RushCore.getInstance().delete(this, callback);
+    }
+
     @Override
-    public String getId() { return RushCore.getInstance().getId(this); }
+    public String getId() {
+        return RushCore.getInstance().getId(this);
+    }
 }
